requirements.txt
tensorflow==2.11.0
transformers==4.27.1
scikit-learn
matplotlib
numpy
pandas
argparse
python-Levenshtein
nltk
